ssu module (#57)
* interface for smart storage unit module

* adding template for tests

* deplouyable and inventory implementation

* - adding erc721token in deployable module
- fixing tests
- function to create meta data for the smart deployable

* add smart deployable token deployment

* refactor scripts

* refactoring names to sync with the indexer

* adding scripts for testing

* adding constants

* fixes for deployment scripts

* create character via forwarder

* deleting constants file

* fixes for end to end test scripts

* adding createMint script

+++ mud-contracts/frontier-world/src/modules/entity-record/EntityRecordLib.sol
@@ -26,7 +26,7 @@ library EntityRecordLib {
     World memory world,
     uint256 entityId,
     uint256 itemId,
-    uint8 typeId,
+    uint256 typeId,
     uint256 volume
   ) internal {
     world.iface.call(
@@ -61,33 +61,18 @@ library EntityRecordLib {
     );
   }
 
-  function setName(
-    World memory world,
-    uint256 entityId,
-    string memory name
-  ) internal {
-    world.iface.call(
-      world.namespace.entityRecordSystemId(),
-      abi.encodeCall(IEntityRecord.setName, (entityId, name))
-    );
+  function setName(World memory world, uint256 entityId, string memory name) internal {
+    world.iface.call(world.namespace.entityRecordSystemId(), abi.encodeCall(IEntityRecord.setName, (entityId, name)));
   }
 
-  function setDappURL(
-    World memory world,
-    uint256 entityId,
-    string memory dappURL
-  ) internal {
+  function setDappURL(World memory world, uint256 entityId, string memory dappURL) internal {
     world.iface.call(
       world.namespace.entityRecordSystemId(),
       abi.encodeCall(IEntityRecord.setDappURL, (entityId, dappURL))
     );
   }
 
-    function setDescription(
-    World memory world,
-    uint256 entityId,
-    string memory description
-  ) internal {
+  function setDescription(World memory world, uint256 entityId, string memory description) internal {
     world.iface.call(
       world.namespace.entityRecordSystemId(),
       abi.encodeCall(IEntityRecord.setDescription, (entityId, description))
