+++ mud-contracts/frontier-world/src/modules/entity-record/interfaces/IEntityRecord.sol
@@ -5,7 +5,7 @@ pragma solidity >=0.8.21;
  * @title IEntityRecord system
  */
 interface IEntityRecord {
-  function createEntityRecord(uint256 entityId, uint256 itemId, uint8 typeId, uint256 volume) external;
+  function createEntityRecord(uint256 entityId, uint256 itemId, uint256 typeId, uint256 volume) external;
 
   function createEntityRecordOffchain(
     uint256 entityId,
@@ -21,18 +21,9 @@ interface IEntityRecord {
     string memory description
   ) external;
 
-  function setName(
-    uint256 entityId,
-    string memory name
-  ) external;
+  function setName(uint256 entityId, string memory name) external;
 
-  function setDappURL(
-    uint256 entityId,
-    string memory dappURL
-  ) external;
+  function setDappURL(uint256 entityId, string memory dappURL) external;
 
-  function setDescription(
-    uint256 entityId,
-    string memory description
-  ) external;
+  function setDescription(uint256 entityId, string memory description) external;
 }
